Moved Turn IO files into a well-named folder.
