+++ src/main/java/io/turn_io/TurnPresenter.java
@@ -1,4 +1,4 @@
-package CoreEntities.IO;
+package io.turn_io;
 
 public class TurnPresenter {
